ResourceParamCreator: matched dumpUserBin_
Other changes:
- Reverted Event::calc and BlendContainer::calc to WIP because of sudden mismatch
- Fixed some variables needing to be signed

+++ include/xlink2/xlink2UserInstance.h
@@ -98,7 +98,7 @@ public:
     
     void freeEventIfFadeOrKillCalled();
 
-    u64 getCurrentResActionIdx(s32) const;
+    s32 getCurrentResActionIdx(s32) const;
 
     bool searchAsset(Locator*, char const*);
     bool searchAsset(Locator*, u32);
